refactor: 중복되는 코드를 추상 클래스를 두어 중복을 제거
