+++ src/main/java/step4/domain/state/Finished.java
@@ -1,6 +1,14 @@
 package step4.domain.state;
 
-abstract class Finished implements State {
+import step4.domain.Score;
+
+public abstract class Finished implements State {
+    private Score score;
+
+    public Finished(int score, int left) {
+        this.score = new Score(10, 2);
+    }
+
     @Override
     public State throwBowl(int fallenPins) {
         return null;
@@ -9,5 +17,15 @@ abstract class Finished implements State {
     @Override
     public boolean isFinish() {
         return true;
+    }
+
+    @Override
+    public int getScore() {
+        return score.getScore();
+    }
+
+    @Override
+    public Score score() {
+        return score;
     }
 }
